Added tests and fixed int rounding bug

+++ CrossPlatformPokemonTD/RangeCircle.cs
@@ -6,7 +6,7 @@ namespace PokemonTDCore
 {
     public class RangeCircle : DrawableGameObject
     {
-        public RangeCircle(TextureHelper textureHelper, int range, Vector2 position) : base(new Vector2(position.X - range, position.Y - range), textureHelper.RangeCircle, new Vector2(range * 2, range * 2))
+        public RangeCircle(TextureHelper textureHelper, float range, Vector2 position) : base(new Vector2(position.X - range, position.Y - range), textureHelper.RangeCircle, new Vector2(range * 2, range * 2))
         {
         }
